Rename folder for topic feature.
